Set home directory to be world-readable
When use this image with unpacked.cern.ch (https://github.com/cvmfs/images-unpacked.cern.ch/pull/29), the files are unpacked without uidmap. This causes the runner home directory to only be readable by the runner user (uid 1001 at the time of writing). We want to use this image in a rootless environment. This PR makes it so that the runner home directory is world-readable. This is kinda hacky, but worth experimenting with.